FirebaseViewModel: Don't Delete Image from cloud for Forwarded Messages

+++ app/src/main/java/com/example/swiftwave/ui/viewmodels/FirebaseViewModel.kt
@@ -445,9 +445,11 @@ class FirebaseViewModel: ViewModel() {
                     .set(it)
             }
 
-            messageData.image?.let { imageUrl ->
-                val storageRef = Firebase.storage.getReferenceFromUrl(imageUrl)
-                storageRef.delete()
+            if(messageData.isForwarded == false){
+                messageData.image?.let { imageUrl ->
+                    val storageRef = Firebase.storage.getReferenceFromUrl(imageUrl)
+                    storageRef.delete()
+                }
             }
         }
     }
